+++ Sources/PostLimit/PostLimitNotify.php
@@ -1,20 +1,15 @@
 <?php
 
-namespace tasks;
-
-use PostLimit\PostLimitRepository;
-use PostLimit\PostLimitService;
+namespace PostLimit;
 
 class PostLimitNotify extends \SMF_BackgroundTask
 {
     protected PostLimitRepository $repository;
 
-    public function __construct()
-    {
-        $this->repository = new PostLimitRepository();
-    }
     public function execute(): bool
     {
+        $this->repository = new PostLimitRepository();
+
         $this->repository->insertAlert($this->_details);
 
         return true;
